1. RESTFul. Описание архитектуры [#6884 #451886]

+++ src/main/java/ru/job4j/persons/domain/Person.java
@@ -1,8 +1,7 @@
-/**
-
 package ru.job4j.persons.domain;
 
 import lombok.*;
+
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
@@ -17,13 +16,11 @@ import javax.persistence.Id;
 public class Person {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private long id;
     private String login;
     private String password;
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 }
-
- */
