+++ src/main/java/ru/job4j/persons/repository/PersonRepository.java
@@ -1,13 +1,6 @@
-/**
- *
-
 package ru.job4j.persons.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.job4j.persons.domain.Person;
 
-public interface PersonRepository extends JpaRepository<Person, Integer> {
-
-}
-
- */
+public interface PersonRepository extends JpaRepository<Person, Long> { }
